RefactoringHelper-Class that contains methods that are used in multiple refactorings to reduce duplicate code. Added more specific error handling and logging some exceptions when javaparser tries to resolve superclasses from external projects
