+++ src/main/java/de/refactoringBot/refactoring/RefactoringHelper.java
@@ -1,0 +1,310 @@
+package de.refactoringBot.refactoring;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.nio.file.InvalidPathException;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.github.javaparser.JavaParser;
+import com.github.javaparser.ast.CompilationUnit;
+import com.github.javaparser.ast.PackageDeclaration;
+import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
+import com.github.javaparser.ast.body.MethodDeclaration;
+import com.github.javaparser.ast.expr.MethodCallExpr;
+import com.github.javaparser.printer.lexicalpreservation.LexicalPreservingPrinter;
+import com.github.javaparser.resolution.UnsolvedSymbolException;
+import com.github.javaparser.resolution.declarations.ResolvedMethodDeclaration;
+import com.github.javaparser.resolution.declarations.ResolvedReferenceTypeDeclaration;
+import com.github.javaparser.resolution.types.ResolvedReferenceType;
+
+import de.refactoringBot.model.exceptions.BotRefactoringException;
+import de.refactoringBot.model.javaparser.ParserRefactoring;
+
+public class RefactoringHelper {
+	
+	// Logger
+	private static final Logger logger = LoggerFactory.getLogger(RefactoringPicker.class);
+
+	/**
+	 * This method returns all root-folders of java files (like the src folder or
+	 * the src/main/java folder from maven projects)
+	 * 
+	 * @param allJavaFiles
+	 * @param repoFolder
+	 * @return javaRoots
+	 * @throws FileNotFoundException
+	 */
+	public List<String> findJavaRoots(List<String> allJavaFiles, String repoFolder) throws FileNotFoundException {
+
+		// Init roots list
+		List<String> javaRoots = new ArrayList<>();
+
+		for (String javaFile : allJavaFiles) {
+			// parse a file
+			FileInputStream filepath = new FileInputStream(javaFile);
+			CompilationUnit compilationUnit = LexicalPreservingPrinter.setup(JavaParser.parse(filepath));
+
+			// Get all Classes
+			List<PackageDeclaration> packageDeclarations = compilationUnit.findAll(PackageDeclaration.class);
+
+			// If javafile has no package
+			if (packageDeclarations.isEmpty()) {
+				// Get javafile
+				File rootlessFile = new File(javaFile);
+				// Add parent of file to root
+				if (!javaRoots.contains(rootlessFile.getParentFile().getAbsoluteFile().getAbsolutePath())) {
+					javaRoots.add(rootlessFile.getParentFile().getAbsolutePath());
+				}
+			} else {
+				// Only 1 package declaration for each file
+				PackageDeclaration packageDeclaration = packageDeclarations.get(0);
+				String rootPackage = null;
+
+				if (packageDeclaration.getNameAsString().split("\\.").length == 1) {
+					rootPackage = packageDeclaration.getNameAsString();
+				} else {
+					rootPackage = packageDeclaration.getNameAsString().split("\\.")[0];
+				}
+
+				// Get javafile
+				File currentFile = new File(javaFile);
+
+				// Until finding the root package
+				while (!currentFile.isDirectory() || !currentFile.getName().equals(rootPackage)) {
+					currentFile = currentFile.getParentFile();
+				}
+
+				// Add parent of rootPackage as java root
+				if (!javaRoots.contains(currentFile.getParentFile().getAbsoluteFile().getAbsolutePath())) {
+					javaRoots.add(currentFile.getParentFile().getAbsolutePath());
+				}
+			}
+
+		}
+
+		return javaRoots;
+	}
+	
+	/**
+	 * This method scanns all java files for method calls that need to be renamed.
+	 * 
+	 * @param refactoring
+	 * @param allJavaFiles
+	 * @return
+	 * @throws FileNotFoundException
+	 */
+	public ParserRefactoring findMethods(ParserRefactoring refactoring, List<String> allJavaFiles,
+			String methodSignature) throws FileNotFoundException {
+
+		// Iterate all Java-Files
+		for (String javaFile : allJavaFiles) {
+			// parse a file
+			FileInputStream filepath = new FileInputStream(javaFile);
+			CompilationUnit compilationUnit = LexicalPreservingPrinter.setup(JavaParser.parse(filepath));
+
+			// Get all Classes
+			List<ClassOrInterfaceDeclaration> classes = compilationUnit.findAll(ClassOrInterfaceDeclaration.class);
+			// Iterate all Classes
+			for (ClassOrInterfaceDeclaration currentClass : classes) {
+
+				// If class is sub/or superclass
+				if (refactoring.getClasses().contains(currentClass.resolve().getQualifiedName())) {
+					// Get all methods
+					List<MethodDeclaration> methods = currentClass.getMethods();
+					// Search methods
+					for (MethodDeclaration method : methods) {
+						if (method.getSignature().asString().equals(methodSignature)) {
+							refactoring.addMethod(method);
+							refactoring.addMethodSignature(getFullMethodSignature(method));
+							if (!refactoring.getJavaFiles().contains(javaFile)) {
+								refactoring.addJavaFile(javaFile);
+							}
+						}
+					}
+				}
+			}
+		}
+
+		return refactoring;
+	}
+	
+	/**
+	 * This method scanns all java files for methods that need to be renamed.
+	 * 
+	 * @param refactoring
+	 * @param allJavaFiles
+	 * @return
+	 * @throws FileNotFoundException
+	 */
+	public ParserRefactoring findMethodCalls(ParserRefactoring refactoring, List<String> allJavaFiles)
+			throws FileNotFoundException {
+
+		for (String javaFile : allJavaFiles) {
+			// parse a file
+			FileInputStream filepath = new FileInputStream(javaFile);
+			CompilationUnit compilationUnit = LexicalPreservingPrinter.setup(JavaParser.parse(filepath));
+
+			List<MethodCallExpr> methodCalls = compilationUnit.findAll(MethodCallExpr.class);
+
+			for (MethodCallExpr methodCall : methodCalls) {
+
+				ResolvedMethodDeclaration calledMethod = null;
+				try {
+					calledMethod = methodCall.resolve();
+				} catch (Exception e) {
+					logger.debug(e.getMessage());
+					continue;
+				}
+
+				if (refactoring.getMethodSignatures().contains(calledMethod.getQualifiedSignature())) {
+					refactoring.addMethodCall(methodCall);
+
+					if (!refactoring.getJavaFiles().contains(javaFile)) {
+						refactoring.addJavaFile(javaFile);
+					}
+				}
+			}
+
+		}
+
+		return refactoring;
+	}
+	
+	/**
+	 * This method scanns all java files for classes that could be subclasses in our
+	 * AST-Tree.
+	 * 
+	 * @param refactoring
+	 * @param allJavaFiles
+	 * @return refactoring
+	 * @throws FileNotFoundException
+	 * @throws BotRefactoringException
+	 */
+	public ParserRefactoring addSubClasses(ParserRefactoring refactoring, List<String> allJavaFiles)
+			throws FileNotFoundException {
+
+		// Search all Java-Files
+		for (String javaFile : allJavaFiles) {
+			// parse a file
+			FileInputStream filepath = new FileInputStream(javaFile);
+			CompilationUnit compilationUnit = LexicalPreservingPrinter.setup(JavaParser.parse(filepath));
+
+			// Get all Classes
+			List<ClassOrInterfaceDeclaration> classes = compilationUnit.findAll(ClassOrInterfaceDeclaration.class);
+
+			// Search all Classes
+			for (ClassOrInterfaceDeclaration currentClass : classes) {
+				List<String> classAncestors = new ArrayList<String>();
+				boolean isSubClass = false;
+				boolean hasExternalDep = false;
+
+				// Get all Super-Classes
+				List<ResolvedReferenceType> ancestors = null;
+				try {
+					ancestors = currentClass.resolve().getAllAncestors();
+				} catch (InvalidPathException | UnsolvedSymbolException i) {
+					ancestors = getAllAncestors(currentClass.resolve());
+					hasExternalDep = true;
+				} catch (Exception e) {
+					logger.debug(e.getMessage());
+					continue;
+				}
+
+				for (ResolvedReferenceType ancestor : ancestors) {
+					// Collect all Signatures of Super-Classes
+					if (!ancestor.getQualifiedName().equals("java.lang.Object")) {
+						classAncestors.add(ancestor.getQualifiedName());
+					}
+
+					// If Super-Class belongs to AST-Tree
+					if (refactoring.getClasses().contains(ancestor.getQualifiedName())
+							&& !refactoring.getClasses().contains(currentClass.resolve().getQualifiedName())) {
+						// Add class to AST-Tree
+						refactoring.getClasses().add(currentClass.resolve().getQualifiedName());
+						isSubClass = true;
+					}
+				}
+
+				// Add all Super-Classes of Class to AST-Tree
+				if (isSubClass) {
+					for (String classAncestor : classAncestors) {
+						if (!refactoring.getClasses().contains(classAncestor)) {
+							refactoring.getClasses().add(classAncestor);
+						}
+					}
+					// Add warning
+					if (hasExternalDep) {
+						refactoring.setWarning(
+								" Refactored classes might extend/implement external project! Check if overriden method was NOT renamed!");
+					}
+				}
+			}
+		}
+
+		return refactoring;
+	}
+	
+	/**
+	 * This method returns the global signature of a method as a string.
+	 * 
+	 * @param methodDeclaration
+	 * @param position
+	 * @return
+	 */
+	public String getFullMethodSignature(MethodDeclaration methodDeclaration) {
+		ResolvedMethodDeclaration resolvedMethod = methodDeclaration.resolve();
+		return resolvedMethod.getQualifiedSignature();
+	}
+	
+	/**
+	 * This method returns the local signature of a method as a string.
+	 * 
+	 * @param methodDeclaration
+	 * @param position
+	 * @return
+	 */
+	public String getMethodDeclarationAsString(MethodDeclaration methodDeclaration, Integer position) {
+		// If method is at the refactored position
+		if (methodDeclaration.getName().getBegin().isPresent()) {
+			if (position == methodDeclaration.getName().getBegin().get().line) {
+				return methodDeclaration.getSignature().asString();
+			}
+		}
+		return null;
+	}
+	
+	/**
+	 * This method gets all direct and indirect Ancestors of a given class if
+	 * possible. (If ancestor is not a external dependency for example)
+	 * 
+	 * @param currentClass
+	 * @return ancestors
+	 */
+	public List<ResolvedReferenceType> getAllAncestors(ResolvedReferenceTypeDeclaration currentClass) {
+		// Init ancestor list
+		List<ResolvedReferenceType> ancestors = new ArrayList<>();
+
+		// Check class
+		if (!(Object.class.getCanonicalName().equals(currentClass.getQualifiedName()))) {
+			// Get all direct ancestors that can be resolved
+			for (ResolvedReferenceType ancestor : currentClass.getAncestors(true)) {
+				// Add them to list
+				ancestors.add(ancestor);
+				// Get indirect ancestors recursively
+				for (ResolvedReferenceType inheritedAncestor : getAllAncestors(ancestor.getTypeDeclaration())) {
+					if (!ancestors.contains(inheritedAncestor)) {
+						ancestors.add(inheritedAncestor);
+					}
+				}
+			}
+		}
+
+		return ancestors;
+	}
+
+}
